requirements.txt
Flask
joblib
numpy
transformers
rembg
Pillow
matplotlib
scikit-learn
gunicorn
tensorflow
tf-keras